Allow credit card numbers to be pasted with `now cc add` (#656)
This fixes a bug where, when pasting a credit card, it would print the
number correctly, but also append too much placeholder. When hitting
enter, backspace, or another number, it would begin printing (and
formatting) `undefined`. This adds a special case where the CC will be
formatted to account for spaces when comparing against the length of the
existing value or placeholder.