Remove whois check when using domains add command
When a domain is added with `now domains add` command it's often
because the user needs a working Zone file before changing the NS
records at the registrar or due to issues with the whois service.
Therefore the command is almost always run with `-f` flag.

Remove the `--force` option and always add send the domain add
request without client side verification.